backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
twilio==8.10.0
openai==1.3.0
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
sqlitecloud==0.0.84
